requirements.txt
asgiref==3.3.1
attrs==20.3.0
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
drf-spectacular==0.13.2
inflection==0.5.1
jsonschema==3.2.0
Pillow==8.1.0
pyrsistent==0.17.3
pytz==2021.1
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
